Added section on nodes, revised steps
I added a general explanation of what nodes are and what workflows are,
referring back to the diagram on the Nipype and Neuroimaging page.  I
edited the breakdown of the steps, too, to be a little more verbose, which
I think is good for the beginners.  I hope I did not get anything too
wrong.